// src/usecase/customer/create/create.customer.dto.ts
export interface InputCreateCustomer {
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    }
}

export interface OutputCreateCustomer {
    id: string;
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    }
}